+++ go.mod
@@ -1,14 +1,16 @@
 module github.com/mosajjal/sniproxy
 
-go 1.18
+go 1.19
 
 require (
 	github.com/golang-collections/collections v0.0.0-20130729185459-604e922904d3
 	github.com/miekg/dns v1.1.50
 	github.com/mosajjal/dnsclient v0.1.0
 	github.com/mosajjal/doqd v0.0.0-20221017212049-9745a8eb6912
+	github.com/oschwald/maxminddb-golang v1.10.0
 	github.com/sirupsen/logrus v1.9.0
 	github.com/spf13/pflag v1.0.5
+	golang.org/x/exp v0.0.0-20221028150844-83b7d23a625f
 )
 
 require (
@@ -28,7 +30,6 @@ require (
 	github.com/prometheus/common v0.37.0 // indirect
 	github.com/prometheus/procfs v0.8.0 // indirect
 	golang.org/x/crypto v0.1.0 // indirect
-	golang.org/x/exp v0.0.0-20221025133541-111beb427cde // indirect
 	golang.org/x/mod v0.6.0 // indirect
 	golang.org/x/net v0.1.0 // indirect
 	golang.org/x/sys v0.1.0 // indirect
